backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1
django-ratelimit==4.1.0
Pillow
python-decouple==3.8
psycopg2-binary
whitenoise==6.6.0
drf-spectacular==0.27.2
channels==4.1.0
channels-redis==4.2.0
redis==5.0.4
gunicorn==22.0.0
daphne==4.1.2
django-storages==1.14.3
boto3==1.34.69
sentry-sdk==2.3.1
django-extensions==3.2.3
